Require auto_index attribute to enable automatic index assigment
Automatically assigning indices based on the order of the fields is
error prone, for example when used with feature flags, when fields are
skipped or when the fields of the struct are changed or reordered.
With this patch, we require an explicit opt-in to this potentially
problematic behavior.  See also:
     https://github.com/trussed-dev/serde-indexed/issues/17